+++ src/dipdup/http.py
@@ -6,6 +6,7 @@ import platform
 from abc import ABC
 from contextlib import suppress
 from http import HTTPStatus
+from typing import Any
 from typing import Mapping
 from typing import Optional
 from typing import Tuple
@@ -43,9 +44,16 @@ class HTTPGateway(ABC):
         """Close underlying aiohttp session"""
         await self._http.__aexit__(exc_type, exc, tb)
 
-    async def close_session(self) -> None:
-        """Close aiohttp session"""
-        await self._http.close_session()
+    async def request(
+        self,
+        method: str,
+        url: str,
+        cache: bool = False,
+        weight: int = 1,
+        **kwargs,
+    ) -> Any:
+        """Send arbitrary HTTP request"""
+        return await self._http.request(method, url, cache, weight, **kwargs)
 
     def set_user_agent(self, *args: str) -> None:
         """Add list of arguments to User-Agent header"""
@@ -163,7 +171,14 @@ class _HTTPGateway:
         ) as response:
             return await response.json()
 
-    async def request(self, method: str, url: str, cache: bool = False, weight: int = 1, **kwargs):
+    async def request(
+        self,
+        method: str,
+        url: str,
+        cache: bool = False,
+        weight: int = 1,
+        **kwargs,
+    ) -> Any:
         """Perform an HTTP request.
 
         Check for parameters in cache, if not found, perform retried request and cache result.
@@ -178,10 +193,6 @@ class _HTTPGateway:
                 return response  # noqa: R504
         else:
             return await self._retry_request(method, url, weight, **kwargs)
-
-    async def close_session(self) -> None:
-        """Close aiohttp session"""
-        await self._session.close()
 
     def set_user_agent(self, *args: str) -> None:
         """Add list of arguments to User-Agent header"""