+++ src/dipdup/http.py
@@ -6,6 +6,7 @@ import platform
 from abc import ABC
 from contextlib import suppress
 from http import HTTPStatus
+from json import JSONDecodeError
 from typing import Mapping
 from typing import Optional
 from typing import Tuple
@@ -161,7 +162,10 @@ class _HTTPGateway:
             raise_for_status=True,
             **kwargs,
         ) as response:
-            return await response.json()
+            try:
+                return await response.json()
+            except (JSONDecodeError, aiohttp.ContentTypeError):
+                return await response.read()
 
     async def request(self, method: str, url: str, cache: bool = False, weight: int = 1, **kwargs):
         """Perform an HTTP request.