+++ src/dipdup/config.py
@@ -956,6 +956,7 @@ default_hooks = {
 @dataclass
 class AdvancedConfig:
     reindex: Dict[ReindexingReasonC, ReindexingAction] = field(default_factory=dict)
+    scheduler: Optional[Dict[str, Any]] = None
     oneshot: bool = False
     postpone_jobs: bool = False
     early_realtime: bool = False