+++ CHANGELOG.md
@@ -2,11 +2,22 @@
 
 Please use [this](https://docs.gitlab.com/ee/development/changelog.html) document as guidelines to keep a changelog.
 
+## [unreleased]
+
+### Added
+
+* context: Added `ctx.get_<kind>_datasource` helpers to avoid type casting.
+* ipfs: Added `ipfs` datasource to download JSON and binary data from IPFS.
+
+### Deprecated
+
+* bcd: Added deprecation notice.
+
 ## 4.1.2 - 2022-01-27
 
 ### Added
 
-* cli: Added `schema wipe --force` argument to skip confirmation promt.
+* cli: Added `schema wipe --force` argument to skip confirmation prompt.
 
 ### Fixed
 
@@ -21,7 +32,7 @@ Please use [this](https://docs.gitlab.com/ee/development/changelog.html) documen
 * cli: Fixed stacktraces missing on exception.
 * cli: Fixed wrapping `OSError` with `ConfigurationError` during config loading.
 * hasura: Fixed printing help messages on `HasuraError`.
-* hasura: Preserve list of sources in Hasura Cloud environments.
+* hasura: Preserve a list of sources in Hasura Cloud environments.
 * hasura: Fixed `HasuraConfig.source` config option.
 
 ### Changed